+++ src/features/mypage/my-recipes/types/index.ts
@@ -1,0 +1,6 @@
+export interface RecipeInfoData {
+  recipeId: number;
+  foodName: string;
+  content: string;
+  category: number;
+}
