--- src/types/findModal.ts
@@ -1,5 +1,0 @@
-export interface FormValues {
-  email: string;
-  emailConfirm?: string;
-  id?: string;
-}
