--- src/types/login.ts
@@ -1,4 +1,0 @@
-export interface FormValues {
-  id: string;
-  password: string;
-}
